Adding Application Signals Contract Test for MySQL
