+++ appsignals-tests/contract-tests/build.gradle.kts
@@ -58,6 +58,8 @@ dependencies {
   implementation(project(":appsignals-tests:images:grpc:grpc-base"))
   testImplementation("org.testcontainers:kafka:1.19.3")
   testImplementation("org.testcontainers:postgresql:1.19.3")
+  testImplementation("org.testcontainers:mysql:1.19.8")
+  testImplementation("com.mysql:mysql-connector-j:8.4.0")
 }
 
 project.evaluationDependsOn(":otelagent")
